integration_tests/requirements.txt
behave==1.2.6
behave-webdriver==0.3.1
webdriver-manager==3.8.5
msedge-selenium-tools==3.141.4
